Clean settings components
