+++ src/settings/components/form/PropertiesForm.scss
