+++ src/settings/components/ui/AddButton.scss
