+++ src/settings/components/ui/DeleteButton.scss
