+++ src/settings/components/ui/Input.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './input.scss';
+import './Input.scss';
 
 class Input extends React.Component {
